add 3 layered pho bowl example

+++ src/main/java/wordcloud/font/scale/FontScalar.java
@@ -4,5 +4,5 @@ package wordcloud.font.scale;
  * Created by kenny on 6/30/14.
  */
 public interface FontScalar {
-    int scale(int n, int minValue, int maxValue);
+    float scale(int n, int minValue, int maxValue);
 }
